factor the formcontext module out completely, moving all article related forms into the forms module
