+++ doajtest/unit/apiv2_tests/test_apiv2_crud_application.py
@@ -7,7 +7,6 @@ from doajtest.helpers import DoajTestCase
 from portality.api.common import Api401Error, Api400Error, Api404Error, Api403Error
 from portality.api.v2.crud.applications import ApplicationsCrudApi
 from portality.api.v2.data_objects.application import IncomingApplication, OutgoingApplication
-from portality.formcontext import formcontext
 from portality.lib.dataobj import ScriptTagFoundException
 from portality.lib.seamless import SeamlessException
 from portality.forms.application_processors import ApplicationProcessor
@@ -23,12 +22,12 @@ def mock_custom_validate_always_pass(self, *args, **kwargs):
 class TestCrudApplication(DoajTestCase):
 
     def setUp(self):
-        self.old_finalise = formcontext.FormContext.finalise
+        self.old_finalise = ApplicationProcessor.finalise
         self.old_custom_validate = IncomingApplication.custom_validate
         super(TestCrudApplication, self).setUp()
 
     def tearDown(self):
-        formcontext.FormContext.finalise = self.old_finalise
+        ApplicationProcessor.finalise = self.old_finalise
         IncomingApplication.custom_validate = self.old_custom_validate
         super(TestCrudApplication, self).tearDown()
