+++ docs/featuremap/types.csv
@@ -15,6 +15,9 @@ ExternalService
 Feature
 Form
 FormContext
+FormField
+FormValidator
+FormWidget
 Fragment
 Framework
 Library
