+++ portality/view/forms.py
@@ -9,6 +9,7 @@ import re
 from wtforms import Form, validators
 from wtforms import StringField, TextAreaField, HiddenField
 
+import regex
 from portality import models
 from portality.forms.validate import OptionalIf, MaxLen
 
@@ -73,7 +74,6 @@ class EditorGroupForm(Form):
 ## Continuations Forms
 ##########################################################################
 
-ISSN_REGEX = re.compile(r'^\d{4}-\d{3}(\d|X){1}$', re.IGNORECASE)
 ISSN_ERROR = 'An ISSN or EISSN should be 7 or 8 digits long, separated by a dash, e.g. 1234-5678. If it is 7 digits long, it must end with the letter X (e.g. 1234-567X).'
 
 
@@ -82,11 +82,11 @@ class MakeContinuation(Form):
     title = StringField('Journal Title', [validators.DataRequired()])
 
     pissn = StringField('Journal ISSN (print version)',
-        [OptionalIf('eissn'), validators.Regexp(regex=ISSN_REGEX, message=ISSN_ERROR)],
+        [OptionalIf('eissn'), validators.Regexp(regex=regex.ISSN_COMPILED, message=ISSN_ERROR)],
         description='Only provide the print ISSN if your journal has one, otherwise leave this field blank. Write the ISSN with the hyphen "-" e.g. 1234-4321.',
     )
     eissn = StringField('Journal ISSN (online version)',
-        [OptionalIf('pissn'), validators.Regexp(regex=ISSN_REGEX, message=ISSN_ERROR)],
+        [OptionalIf('pissn'), validators.Regexp(regex=regex.ISSN_COMPILED, message=ISSN_ERROR)],
         description='Cannot be the same as the P-ISSN. Write the EISSN with the hyphen "-" e.g. 1234-4321.',
     )
